# docs/swagger.yaml
definitions:
  internalErrors.RestErr:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  users.User:
    properties:
      date_created:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      password:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
  /internal/users/search:
    get:
      description: search a user based on status
      parameters:
      - description: status
        enum:
        - active
        in: query
        name: enumstring
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
      summary: Search user by status
      tags:
      - users
  /ping:
    get:
      description: return a pong in order to test health
      responses:
        "200":
          description: pong
      summary: Return pong
      tags:
      - ping
  /users:
    post:
      description: create a new user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      description: delete user from db
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: status:deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
      summary: Delete user
      tags:
      - users
    get:
      description: get user info
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
      summary: Get user
      tags:
      - users
    patch:
      description: update user info
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
      summary: Update user
      tags:
      - users
    put:
      description: update user info
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internalErrors.RestErr'
      summary: Update user
      tags:
      - users
swagger: "2.0"